AI Attacking Changes
When Desirable Block is zero (which is now an unsigned 32-bit integer), the Health considered to be 'Quite Healthy' has been increased from 14 to 16.

Additionally, if the block-distance between the AI Warrior and the Nearest Human is greater than 10, the AI will compute a flanking manoeuvre. Otherwise, it will continue as before.

AWarrior::Flank() now needs an ABlock* to calculate a flank path.

If a Warrior is unable to AWarrior::FindNearestAffiliation(), then they will simply go towards the corresponding concentration.

NOTE: A restructuring of the AI when DesirableBlocks == 0 is imminent.